EODHP-1117 - update dockerfile
